requirements.txt
python-dotenv
requests
gnews
yfinance
twikit
selenium
bs4
webdriver-manager
duckduckgo_search
httpx
tldextract
thefuzz
pandas
ta
fastapi
uvicorn